Cleanup probing
Since sif_platform_rescan_card() is just an empty stub we need (hack)
broken-cd in the devicetree with upstream kernels to redetect the card
after the initial firmware load. When leaving the redection to the
mmc-core like this, the probe function will get called again
automatically when the mmc-core does the redetection.

This makes the entire unregister / register and semaphore dance
unnecessary, therefor this commit removes it.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>