Add support for mmc_force_detect_change()
I've send a patch to the upstream kernel which adds a
mmc_force_detect_change() function to the upstream kernel which
allows things to work with non-removable in the devicetree instead
of using broken-cd as a hack.

This commit adds support for it, as well as adding a stub implementation
to use until this gets merged upstream.

Note that the stub does *not* fix the issue of the mmc core powering
down the device between the first and second probe, so until this is
merged upstream the regulators will still need regulator-always-on.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>